+++ Reversed Binary Numbers/main.cpp
@@ -2,7 +2,7 @@
 #include <string>
 #include <cstdlib>
 #include <vector>
-#include <math.h> 
+#include <math.h>
 #include <algorithm>
 
 int main(int argc, char **argv)
@@ -19,19 +19,19 @@ int main(int argc, char **argv)
     binary.push_back(num % 2);
     num = num / 2;
   }
-  if (num) 
+  if (num)
   {
     binary.push_back(1);
   }
-  // Now vector binary contains th reversed binary number from binary.start() 
+  // Now vector binary contains th reversed binary number from binary.start()
   // -> binary.end()
   int reversed_num = 0;
-  std::reverse(binary.begin(), binary.end()); 
-  for (int i = 0; i < binary.size(); i++)
+  std::reverse(binary.begin(), binary.end());
+  for (size_t i = 0; i < binary.size(); i++)
   {
     reversed_num += pow(2, i) * binary.at(i);
   }
   std::cout << reversed_num << "\n";
-  
+
   return 0;
 }
